api_yamdb/requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.1
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-debug-toolbar==3.2.4
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==3.9.2
flake8-docstrings==1.6.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
Markdown==3.4.1
mccabe==0.6.1
packaging==23.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.6.0
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
zipp==3.14.0
# доп установлено
djoser
djangorestframework-simplejwt==4.7.2
django-filter-23.1 